Move new files to services/searchengine
