+++ .gitignore
@@ -166,8 +166,8 @@ ReactNativeClient/lib/services/UndoRedoService.js
 ReactNativeClient/lib/ShareExtension.js
 ReactNativeClient/lib/shareHandler.js
 ReactNativeClient/lib/versionInfo.js
-ReactNativeClient/lib/services/filterParser.js
-ReactNativeClient/lib/services/queryBuilder.js
+ReactNativeClient/lib/services/searchengine/filterParser.js
+ReactNativeClient/lib/services/searchengine/queryBuilder.js
 ReactNativeClient/PluginAssetsLoader.js
 ReactNativeClient/setUpQuickActions.js
 # AUTO-GENERATED - EXCLUDED TYPESCRIPT BUILD
