+++ ElectronClient/plugins/GotoAnything.jsx
@@ -2,8 +2,8 @@ const React = require('react');
 const { connect } = require('react-redux');
 const { _ } = require('lib/locale.js');
 const { themeStyle } = require('lib/theme');
-const SearchEngine = require('lib/services/SearchEngine');
 const CommandService = require('lib/services/CommandService').default;
+const SearchEngine = require('lib/services/searchengine/SearchEngine');
 const BaseModel = require('lib/BaseModel');
 const Tag = require('lib/models/Tag');
 const Folder = require('lib/models/Folder');
