+++ ReactNativeClient/lib/services/ResourceService.js
@@ -4,7 +4,7 @@ const Note = require('lib/models/Note');
 const Resource = require('lib/models/Resource');
 const BaseModel = require('lib/BaseModel');
 const BaseService = require('lib/services/BaseService');
-const SearchEngine = require('lib/services/SearchEngine');
+const SearchEngine = require('lib/services/searchengine/SearchEngine');
 const Setting = require('lib/models/Setting');
 const { shim } = require('lib/shim');
 const ItemChangeUtils = require('lib/services/ItemChangeUtils');
