+++ ReactNativeClient/lib/services/searchengine/SearchEngineUtils.js
@@ -1,4 +1,4 @@
-const SearchEngine = require('lib/services/SearchEngine');
+const SearchEngine = require('lib/services/searchengine/SearchEngine');
 const Note = require('lib/models/Note');
 
 class SearchEngineUtils {
