+++ ReactNativeClient/lib/services/searchengine/queryBuilder.ts
