+++ ReactNativeClient/root.js
@@ -61,7 +61,7 @@ const DropdownAlert = require('react-native-dropdownalert').default;
 const ShareExtension = require('lib/ShareExtension.js').default;
 const handleShared = require('lib/shareHandler').default;
 const ResourceFetcher = require('lib/services/ResourceFetcher');
-const SearchEngine = require('lib/services/SearchEngine');
+const SearchEngine = require('lib/services/searchengine/SearchEngine');
 const WelcomeUtils = require('lib/WelcomeUtils');
 const { themeStyle } = require('lib/components/global-style.js');
 const { uuid } = require('lib/uuid.js');
